src/access_intent/intent/semantic_retriever.py
from typing import Dict, Optional, List
import numpy as np


class SemanticRetriever:
    """语义检索器"""
    
    def __init__(self):
        # 简化实现，避免下载模型
        self.intent_labels = ["search", "code", "chat", "task", "question"]
    
    async def search(self, user_input: str) -> Optional[Dict]:
        """语义搜索"""
        # 简化实现，基于关键词匹配
        intent_keywords = {
            "search": ["搜索", "查找", "查询"],
            "code": ["代码", "编程", "写代码"],
            "chat": ["聊天", "对话", "闲聊"],
            "task": ["任务", "工作", "计划"],
            "question": ["问题", "疑问", "怎么", "如何"]
        }
        
        for intent, keywords in intent_keywords.items():
            for keyword in keywords:
                if keyword in user_input:
                    return {
                        "intent_type": intent,
                        "confidence": 0.8,
                        "distance": 0.2
                    }
        
        return None
